Create a dummy libxcb-xlib.so.1
This library used to exist, but was deprecated.
On older systems where it still exists (e.g SUSE 11), we may end up
mixing the system's libxcb-xlib with the libX11 and libxcb provided by
conda.
This can cause missing symbols when trying to run a binary which links
to libxcb.

NOTE:

If you are facing this problem you also need to install conda-forge's
libx11. libx11 was the only real dependee of libxcb-xlib and was modified to
acommodate the libxcb-xlib removal:

https://lists.freedesktop.org/archives/xcb/2009-April/004489.html
https://lists.freedesktop.org/archives/xcb/2009-April/004492.html