Update for conda-build 3.
Start using the newfangled `{{compiler()}}` templates. With the advent of the
concept of the "build prefix", we need to use its path when re-autogening the
configure script since that's where the MinGW pkg-config file lives. Finally,
in the new system, `$VS_MAJOR` seems not to be defined when building anymore,
so it doesn't work as a way to detect that we're being built on Windows.