Add pthread-stubs as a runtime dependency
Without including this, every single package that does a pkg-config check for
the `xcb` module needs to remember to pull in `pthread-stubs` as a dependency:
XCB's pkg-config file requires `pthread-stubs.pc` and pkg-config checks will
break if the latter isn't available. This is admittedly a secondary use case,
but this problem keeps on biting me, and it can cause subtle failures because
problems happen even when a package only depends implicitly on libxcb.

Ideally, instead of this, we'd start having `-devel` packages and/or dev-only
dependencies, but conda-build and conda-forge are a long ways away from that.
The extra package to install is annoying but it contains literally just the
one pkg-config file.